1.通用函数转移至utils.sh 2.完善命令参数解析

+++ deploy.sh
@@ -5,6 +5,8 @@
 # author: caosiyang <csy3228@gmail.com>
 # date: 2013/06/25
 
+test -f host.conf || exit 1
+
 # zip
 test -f myycsb.zip && rm -f myycsb.zip
 zip myycsb.zip -r myycsb -x myycsb/result/*
@@ -13,9 +15,9 @@ zip myycsb.zip -r myycsb -x myycsb/result/*
 while read remotehost
 do
     echo "deploying $remotehost..."
-    ssh -o StrictHostKeyChecking=no root@$remotehost "mkdir -p /home/caosiyang && rm -rf /home/caosiyang/myycsb*" </dev/null
+    ssh -n -o StrictHostKeyChecking=no root@$remotehost "mkdir -p /home/caosiyang && rm -rf /home/caosiyang/myycsb*"
     scp myycsb.zip root@$remotehost:/home/caosiyang
-    ssh -o StrictHostKeyChecking=no root@$remotehost "cd /home/caosiyang && unzip -oq myycsb.zip" </dev/null
+    ssh -n -o StrictHostKeyChecking=no root@$remotehost "cd /home/caosiyang && unzip -oq myycsb.zip"
 done <host.conf
 
 rm -f myycsb.zip
